putting types in their own file
easier for doc and maintaining
