+++ slskd_api/apis/_base.py
