+++ slskd_api/apis/searches.py
@@ -13,54 +13,9 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from .base import *
+from ._base import *
 import uuid
-from typing import Optional, TypedDict
-
-
-class SearchFile(TypedDict):
-    """
-    TypedDict describing a single search file result. Element found in :py:class:`~slskd_api.apis.searches.SearchResponseItem`.
-    """
-    filename: str
-    size: int
-    code: int
-    isLocked: bool
-    extension: str
-    bitRate: int # for lossy format
-    bitDepth: int # for lossless format
-    length: int # in sec
-    sampleRate: int # in Hz
-
-class SearchResponseItem(TypedDict):
-    """
-    TypedDict describing a search response item. Single element of list returned by :py:meth:`~slskd_api.apis.SearchesApi.search_responses`.
-    """
-    fileCount: int
-    files: list[SearchFile]
-    hasFreeUploadSlot: bool
-    lockedFileCount: int
-    lockedFiles: list[SearchFile]
-    queueLength: int
-    token: int
-    uploadSpeed: int
-    username: str
-
-class SearchState(TypedDict):
-    """
-    TypedDict describing search state. Returned by :py:meth:`~slskd_api.apis.SearchesApi.state`.
-    """
-    endedAt: str
-    fileCount: int
-    id: str
-    isComplete: bool
-    lockedFileCount: int
-    responseCount: int
-    responses: list[SearchResponseItem]
-    searchText: str
-    startedAt: str  # ISO date
-    state: str
-    token: int
+from ._types import *
 
 
 class SearchesApi(BaseApi):
@@ -70,7 +25,7 @@ class SearchesApi(BaseApi):
 
     def search_text(self,
                     searchText: str,
-                    id: Optional[str] = None,
+                    id: str = None,
                     fileLimit: int = 10000,
                     filterResponses: bool = True,
                     maximumPeerQueueLength: int = 1000000,
